fix: 패키지 경로 정리 & Book print 메서드 제거
