+++ jaehoon/미션05.md
@@ -18,9 +18,9 @@
 
 ### 풀이
 
-#### section06.example.utility 패키지의 MathHelper 클래스
+#### com.example.utility 패키지의 MathHelper 클래스
 ```java
-package section06.example.utility;
+package study.section06.com.example.utility;
 
 public class MathHelper {
 
@@ -30,11 +30,11 @@ public class MathHelper {
 }
 ```
 
-#### section06.example 패키지의 UtilityTest 클래스 (클라이언트 코드)
+#### com.example.utility 패키지의 클라이언트 코드 - UtilityTest 클래스
 ```java
-package section06.example;
+package study.section06.com.example;
 
-import static com.example.utility.MathHelper.add;
+import static study.section06.com.example.utility.MathHelper.add;
 
 public class UtilityTest {
 
@@ -55,36 +55,28 @@ public class UtilityTest {
 
 ### 풀이
 
-#### section06.library.books 패키지의 Book 클래스
+#### com.library.books 패키지의 Book 클래스
 ```java
-package section06.library.books;
+package section06.com.library.books;
+
 public class Book {
 
   private String title;
 
   private String author;
 
-  private String isBn;
+  private String isbn;
 
   private boolean isCheckedOut;
 
-  public Book(String title, String author, String isBn) {
+  public Book(String title, String author, String isbn) {
     this.title = title;
     this.author = author;
-    this.isBn = isBn;
-    this.isCheckedOut = false; // 기본값은 대출되지 않은 상태
+    this.isbn = isbn;
   }
 
   public String getTitle() {
     return title;
-  }
-
-  public String getAuthor() {
-    return author;
-  }
-
-  public String getIsBn() {
-    return isBn;
   }
 
   public boolean isCheckedOut() {
@@ -99,17 +91,24 @@ public class Book {
     this.isCheckedOut = false;
   }
 
-  public void print() {
-    System.out.println("도서{제목= " + title + ", 저자= " + author + ", 도서고유번호: " + isBn + "}");
+  @Override
+  public String toString() {
+    return "도서{" +
+        "제목='" + title + '\'' +
+        ", 저자='" + author + '\'' +
+        ", 도서고유번호='" + isbn + '\'' +
+        ", 대출유무=" + isCheckedOut +
+        '}';
   }
 }
 ```
 
-#### section06.library.users 패키지의 User 클래스
+#### com.library.users 패키지의 User 클래스
 ```java
-package section06.library.users;
+package section06.com.library.users;
 
-import section06.library.books.Book;
+import section06.com.library.books.Book;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -153,19 +152,19 @@ public class User {
       System.out.println(name + "님은 대출한 책이 없습니다.");
     } else {
       for (Book book : borrowedBooks) {
-        book.print();
+        System.out.println(book);
       }
     }
   }
 }
 ```
 
-#### section06.library 패키지의 LibraryTest 클래스 (클라이언트 코드)
+#### 클라이언트 코드 - LibraryTest 클래스
 ```java
-package section06.library;
+package section06.com.library;
 
-import section06.library.books.Book;
-import section06.library.users.User;
+import section06.com.library.books.Book;
+import section06.com.library.users.User;
 
 public class LibraryTest {
